backend/requirements.txt
    fastapi
    uvicorn[standard]
    docker
    playwright
    google-generativeai